split usdchaser into the chaser itself and its utils
